requirements.txt
spacy>=3.5
transformers>=4.30
sentence-transformers>=2.2.2
nltk>=3.8
rouge-score>=0.1.2
python-docx
tqdm
torch
numpy
pillow
pytesseract
pandas
joblib
scikit-learn>=1.2
colorama>=0.4.6
tabulate>=0.9.0
reedsolo>=1.7.0
